requirements.txt
google-adk
google-generativeai 
google-genai
python-dotenv
pydantic